various small fixes and changes because of underlying submodule (#11)
* bump submodule

* remove OmOffsetSize_t, but some things still don't align fully...

* binary asserts

* better type magix

* fix test

* remove test for debugging

* writer2 usize cleanup

* remove unneeded lifetime

+++ src/io/reader.rs
@@ -159,7 +159,7 @@ impl<Backend: OmFileReaderBackend> OmFileReader<Backend> {
                 ptr.add(2),
                 ptr.add(4),
                 ptr.add(6),
-                self.reader.lut_chunk_element_count as u64,
+                self.reader.lut_chunk_element_count,
                 LEGACY_IO_SIZE_MERGE,
                 LEGACY_IO_SIZE_MAX,
             )
